fix: update the types of Approvals mocks to the service directories
