--- src/pages/board/outlets/financialReporting/Approvals/types.ts
@@ -1,8 +1,0 @@
-export interface approval_by_credit_request_Mock {
-  approval_id: string;
-  credit_request_id: string;
-  approver_id: string;
-  approver_name: string;
-  concept: string;
-  error: boolean;
-}
