+++ src/services/types.ts
@@ -79,3 +79,11 @@ export type {
   DmDecisions,
   PinnedRequest,
 };
+export interface approval_by_credit_request_Mock {
+  approval_id: string;
+  credit_request_id: string;
+  approver_id: string;
+  approver_name: string;
+  concept: string;
+  error: boolean;
+}
